requirements.txt
appnope==0.1.0
Babel==2.8.0
backcall==0.1.0
certifi==2018.8.24
chardet==3.0.4
decorator==4.4.2
defusedxml==0.5.0
Django==2.2
django-allauth==0.37.1
django-cors-headers==2.4.0
django-phonenumber-field==4.0.0
django-rest-auth==0.9.3
djangorestframework==3.11.0
gunicorn==19.7.1
idna==2.7
ipython==7.13.0
ipython-genutils==0.2.0
jedi==0.17.0
oauthlib==2.1.0
parso==0.7.0
pexpect==4.8.0
phonenumbers==8.12.4
pickleshare==0.7.5
Pillow==7.1.2
prompt-toolkit==3.0.5
psycopg2==2.8.4
ptyprocess==0.6.0
Pygments==2.6.1
python3-openid==3.1.0
pytz==2018.5
requests==2.21.0
requests-oauthlib==1.0.0
six==1.11.0
sqlparse==0.3.0
traitlets==4.3.3
urllib3==1.24.2
wcwidth==0.2.3
whitenoise==3.3.1